@@ -1,5 +1,5 @@
-export const SIGNIN_INCORRECT = "Login incorrect";
-export const SIGNIN_INVALID = "Login invalid";
+export const SIGN_IN_INCORRECT = "Login incorrect";
+export const SIGN_IN_INVALID = "Login invalid";
 export const OTP_INCORRECT = "OTP incorrect";
 export const OTP_REQUIRED = "OTP required";
 export const SIGNUP_INVALID = "User is already exist";
